+++ CHANGES.md
@@ -10,3 +10,6 @@
   - バグ修正
 
 ## develop
+
+- [ADD] オブジェクトタグ API (PutObjectTagging / GetObjectTagging / DeleteObjectTagging) を追加する
+  - @voluntas